Extract toggle to another file
